+++ include/D3D12Helper/D3D12Diagnostics/D3D12Diagnostics.hpp
@@ -1,2 +1,5 @@
 #pragma once
 #include <D3D12Helper/D3D12Diagnostics/D3D12Debug.hpp>
+#include <D3D12Helper/D3D12Diagnostics/D3D12DeviceDiagnostics.hpp>
+#include <D3D12Helper/D3D12Diagnostics/D3D12GpuTimer.hpp>
+#include <D3D12Helper/D3D12Diagnostics/D3D12InfoQueue.hpp>